Fix snipMate issue after using pathogen
To understand what caused the trouble you need to know that vim has a
special folder called after (~/.vim/after) to overrule any other
settings.

One of snipMate file is placed on this "after" directory but after using
pathogen, the direcory is not parsed and that's what caused the trouble.

To understand how this fix works, you can read the following thread:
https://github.com/tpope/vim-pathogen/issues/46

Signed-off-by: Mailson Lira <mailson@gmail.com>